Add KKRT and protocol negotiation
KKRTPSI is now the default protocol for match-cli. Protocol
negotiation is also implemented. Both sender and receiver have
a list of preferred protocols in order of desirability. The sender
sends their entire list of protocols to the receiver who compares
with their own list. They search for an intersection, prioritizing
their own protocol desirability. If there is an intersection, they
respond with that as their decision and the sender views this as a
successful negotiation. If there is no intersection, the receiver
sends the default protocol (KKRTPSI) and the sender views this as
an unsuccessful negotiation but will still proceed.

For now, the user can only input a single desired protocol in the
CLI so their preferred protocol list contains a single element. In
the future, multiple inputs could be allowed as the protocol
negotiation already supports that functionality.